+++ lib/test-suite-helpers/make-proxy.d.ts
@@ -1,0 +1,3 @@
+import { Suman } from "../suman";
+import { ITestSuite } from "suman-types/dts/test-suite";
+export declare const makeProxy: (suman: Suman, ctx: ITestSuite) => Function;
