+++ lib/test-suite-helpers/make-test-suite.js
@@ -9,6 +9,7 @@ var _suman = global.__suman = (global.__suman || {});
 var rules = require('../helpers/handle-varargs');
 var constants = require('../../config/suman-constants').constants;
 var test_suite_base_constructor_1 = require("./test-suite-base-constructor");
+var make_proxy_1 = require("./make-proxy");
 var makeStartSuite = require('./make-start-suite').makeStartSuite;
 var make_handle_befores_afters_1 = require("./make-handle-befores-afters");
 var makeNotifyParent = require('./notify-parent-that-child-is-complete').makeNotifyParent;
@@ -34,7 +35,6 @@ exports.makeTestSuiteMaker = function (suman, gracefulExit, blockInjector) {
             this.interface = suman.interface;
             this.desc = this.title = obj.desc;
             var zuite = this;
-            var ctx = this;
             this.resume = function () {
                 var args = Array.from(arguments);
                 process.nextTick(function () {
@@ -49,43 +49,7 @@ exports.makeTestSuiteMaker = function (suman, gracefulExit, blockInjector) {
             var it = make_it_1.makeIt(suman, zuite);
             var afterAllParentHooks = make_after_all_parent_hooks_1.makeAfterAllParentHooks(suman, zuite);
             var describe = make_describe_1.makeDescribe(suman, gracefulExit, TestSuiteMaker, zuite, notifyParent, blockInjector);
-            var getProxy = function (method, rule, props) {
-                return new Proxy(method, {
-                    get: function (target, prop) {
-                        props = props || [];
-                        var hasSkip = false;
-                        var newProps = props.concat(String(prop)).filter(function (v, i, a) {
-                            if (String(v).toLowerCase() === 'skip') {
-                                hasSkip = true;
-                            }
-                            return a.indexOf(v) === i;
-                        })
-                            .sort()
-                            .map(function (v) { return String(v).toLowerCase(); });
-                        if (hasSkip) {
-                            newProps = ['skip'];
-                        }
-                        var cache, cacheId = newProps.join('-');
-                        var fnCache = ctx.testBlockMethodCache.get(method);
-                        if (!fnCache) {
-                            fnCache = {};
-                            ctx.testBlockMethodCache.set(method, fnCache);
-                        }
-                        if (cache = ctx.testBlockMethodCache.get(method)[cacheId]) {
-                            return cache;
-                        }
-                        var fn = function () {
-                            var args = pragmatik.parse(arguments, rule);
-                            newProps.forEach(function (p) {
-                                args[1][p] = true;
-                            });
-                            args[1].__preParsed = true;
-                            return method.apply(ctx, args);
-                        };
-                        return fnCache[cacheId] = getProxy(fn, rule, newProps);
-                    }
-                });
-            };
+            var getProxy = make_proxy_1.makeProxy(suman, zuite);
             this.describe = this.context = this.suite = getProxy(describe, rules.blockSignature);
             this.it = this.test = getProxy(it, rules.testCaseSignature);
             this.inject = getProxy(inject, rules.hookSignature);
