+++ lib/test-suite-methods/make-describe.js
@@ -83,6 +83,9 @@ exports.makeDescribe = function (suman, gracefulExit, TestSuiteMaker, zuite, not
             return;
         }
         var suite = TestSuiteMaker({ desc: desc, title: desc, opts: opts });
+        if (zuite.fixed) {
+            suite.fixed = true;
+        }
         suite.skipped = opts.skip || zuite.skipped;
         if (!suite.only && suman.describeOnlyIsTriggered) {
             suite.skipped = suite.skippedDueToDescribeOnly = true;
