+++ lib/test-suite-methods/make-describe.ts
@@ -134,6 +134,10 @@ export const makeDescribe = function (suman: ISuman, gracefulExit: Function, Tes
     // note: zuite is the parent of suite; aka, suite is the child of zuite
     const suite = TestSuiteMaker({desc, title: desc, opts});
 
+    if(zuite.fixed){
+      suite.fixed = true;
+    }
+
     // if parent is skipped, child is skipped,
     suite.skipped = opts.skip || zuite.skipped;
